Cleaned up folder structure.
