--- Quick Pad/MenuBar.xaml.cs
@@ -1,14 +1,0 @@
-﻿using Windows.UI.Xaml.Controls;
-
-// The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
-
-namespace Quick_Pad
-{
-    public sealed partial class MenuBar : UserControl
-    {
-        public MenuBar()
-        {
-            this.InitializeComponent();
-        }
-    }
-}
