+++ QuickPad.UI/QuickPad.UI.Common/Dialogs/AskToSave.xaml.cs
