+++ QuickPad.UI/QuickPad.UI.Common/Dialogs/GoToLine.xaml.cs
