+++ QuickPad.UI/QuickPad.UI.Common/Helpers/IntCompare.cs
@@ -1,0 +1,12 @@
+﻿namespace QuickPad.UI.Common.Helpers
+{
+    public static class IntCompare
+    {
+        public const string Equal = "=";
+        public const string NotEqual = "!=";
+        public const string Less = "<";
+        public const string More = ">";
+        public const string LessOrEqual = "<=";
+        public const string MoreOrEqual = ">=";
+    }
+}
