+++ QuickPad.UI/QuickPad.UI.Common/QuickPad.UI.Common.csproj
